0.2023.11.07
Change ImageFile property from App::PropertyFile to App::PropertyFileIncluded.  This way when the user uses "other image file" option in the Environment enumeration the image file will be saved with and embedded into the FreeCAD .FCStd file.  Then if the file is shared the other user will have access to the image file for the texture being used.